+++ Runtime/Interfaces/ISchemaProvider.cs
@@ -1,11 +1,6 @@
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Flock.Models;
-
 namespace Flock.Interfaces
 {
-    //TODO add defined types for all of these 
+    //TODO add defined types for all of these
     public enum SchemaTag
     {
         gameplay,
@@ -14,13 +9,5 @@ namespace Flock.Interfaces
         feature,
         achievement,
         empty
-    }
-    public interface ISchemaProvider
-    {
-        //TODO add summaries
-        Task<List<GameConfigSchema>> GetAllSchemasAsync(SchemaTag tag , CancellationToken cancellationToken = default);
-        Task<List<GameConfigSchema>> GetSchemasByVersionAsync(SchemaTag tag , CancellationToken cancellationToken = default);
-        Task<GameConfigSchema> GetSchemaByIdAsync(string schemaId, CancellationToken cancellationToken = default);
-        Task<List<GamePatchSchema>> GetSchemaConfigsAsync(string schemaId, CancellationToken cancellationToken = default);
     }
 }